feat: add database basic methods and model interface
