+++ .gitignore
@@ -17,3 +17,4 @@ mono_crash.*.json
 .vs
 *.user
 Properties/
+*Backup.csproj
