+++ Database/Models/IDatabaseModel.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Tabloulet.Database.Models
+{
+    public interface IDatabaseModel
+    {
+        Guid Id { get; set; }
+    }
+}
