+++ Database/Models/Image.cs
@@ -1,7 +1,7 @@
 ﻿using SQLite;
 using Tabloulet.Helpers;
 
-namespace Tabloulet.Scenes.Database.Models
+namespace Tabloulet.Database.Models
 {
     [Table(Constants.ImageTable)]
     public class Image : Base
