+++ Database/Models/Page.cs
@@ -2,10 +2,10 @@
 using SQLite;
 using Tabloulet.Helpers;
 
-namespace Tabloulet.Scenes.Database.Models
+namespace Tabloulet.Database.Models
 {
     [Table(Constants.PageTable)]
-    public class Page
+    public class Page : IDatabaseModel
     {
         [PrimaryKey]
         public Guid Id { get; set; }
