+++ Scenes/Home/Home.tscn
@@ -1,0 +1,3 @@
+[gd_scene format=3 uid="uid://dxqo13a3yn4q4"]
+
+[node name="Home" type="Node2D"]
