+++ Tabloulet.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="SQLitePCLRaw.lib.e_sqlite3" Version="2.1.2" />
     <PackageReference Include="SQLitePCLRaw.provider.e_sqlite3" Version="2.1.2" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Scenes\" />
+    <Folder Include="Scenes\Home\" />
+  </ItemGroup>
 </Project>
